weights no longer allow nans
